Spliting classes into separate modules for better organization
